Unify/clean up histogram and binned histogram
